Move ts type related functions in ts-type folder
