--- src/transforms/v2-to-v3/ts-type-ref/index.ts
@@ -1,2 +1,0 @@
-export * from "./getV3ClientTypeName";
-export * from "./replaceTSTypeReference";
