+++ src/transforms/v2-to-v3/utils/index.ts
@@ -1,6 +1,4 @@
 export * from "./getV2ClientNewExpression";
-export * from "./getV2ClientTypeNames";
 export * from "./getV2ServiceModulePath";
 export * from "./hasPropertyWithName";
 export * from "./isTypeScriptFile";
-export * from "./isV2ClientInputOutputType";
